+++ passkey-client/src/lib.rs
@@ -21,15 +21,13 @@ use std::{borrow::Cow, fmt::Display};
 
 use ciborium::{cbor, value::Value};
 use coset::{iana::EnumI64, Algorithm};
-use passkey_authenticator::{
-    Authenticator, CredentialStore, DiscoverabilitySupport, UserValidationMethod,
-};
+use passkey_authenticator::{Authenticator, CredentialStore, UserValidationMethod};
 use passkey_types::{
     crypto::sha256,
     ctap2, encoding,
     webauthn::{
-        self, AuthenticatorExtensionsClientOutputs, AuthenticatorSelectionCriteria,
-        CredentialPropertiesOutput, ResidentKeyRequirement, UserVerificationRequirement,
+        self, AuthenticationExtensionsClientOutputs, AuthenticatorSelectionCriteria,
+        ResidentKeyRequirement, UserVerificationRequirement,
     },
     Passkey,
 };
@@ -37,6 +35,7 @@ use serde::Serialize;
 use typeshare::typeshare;
 use url::Url;
 
+mod extensions;
 mod quirks;
 use quirks::QuirkyRp;
 
@@ -259,8 +258,9 @@ where
             .client_data_hash()
             .unwrap_or_else(|| sha256(client_data_json.as_bytes()).to_vec());
 
-        let cred_props_requested =
-            request.extensions.as_ref().and_then(|ext| ext.cred_props) == Some(true);
+        let extension_request = request.extensions.and_then(|e| e.zip_contents());
+
+        let ctap_extensions = self.registration_extension_ctap2_input(extension_request.as_ref());
 
         let rk = self.map_rk(&request.authenticator_selection, &auth_info);
         let uv = request.authenticator_selection.map(|s| s.user_verification)
@@ -277,7 +277,7 @@ where
                 user: request.user,
                 pub_key_cred_params,
                 exclude_list: request.exclude_credentials,
-                extensions: request.extensions,
+                extensions: ctap_extensions,
                 options: ctap2::make_credential::Options { rk, up: true, uv },
                 pin_auth: None,
                 pin_protocol: None,
@@ -321,21 +321,9 @@ where
                 .map_err(|e| WebauthnError::AuthenticatorError(e.into()))?,
         );
 
-        let cred_props = if cred_props_requested {
-            let auth_discoverability = self.authenticator.store().get_info().await.discoverability;
-            let discoverable = match auth_discoverability {
-                DiscoverabilitySupport::Full => rk,
-                DiscoverabilitySupport::OnlyNonDiscoverable => false,
-                DiscoverabilitySupport::ForcedDiscoverable => true,
-            };
-
-            Some(CredentialPropertiesOutput {
-                discoverable: Some(discoverable),
-                authenticator_display_name: self.authenticator.display_name().cloned(),
-            })
-        } else {
-            None
-        };
+        let store_info = self.authenticator.store().get_info().await;
+        let client_extension_results =
+            self.registration_extension_outputs(extension_request.as_ref(), store_info, rk);
 
         let response = webauthn::CreatedPublicKeyCredential {
             id: encoding::base64url(credential_id.credential_id()),
@@ -350,7 +338,7 @@ where
                 transports: auth_info.transports,
             },
             authenticator_attachment: Some(self.authenticator().attachment_type()),
-            client_extension_results: AuthenticatorExtensionsClientOutputs { cred_props },
+            client_extension_results,
         };
 
         // Sanitize output before sending it back to the RP
@@ -398,6 +386,7 @@ where
             .client_data_hash()
             .unwrap_or_else(|| sha256(client_data_json.as_bytes()).to_vec());
 
+        let ctap_extensions = self.auth_extension_ctap2_input(request.extensions.as_ref());
         let rk = false;
         let uv = request.user_verification != UserVerificationRequirement::Discouraged;
 
@@ -407,7 +396,7 @@ where
                 rp_id: rp_id.to_owned(),
                 client_data_hash: client_data_json_hash.into(),
                 allow_list: request.allow_credentials,
-                extensions: request.extensions,
+                extensions: ctap_extensions,
                 options: ctap2::get_assertion::Options { rk, up: true, uv },
                 pin_auth: None,
                 pin_protocol: None,
@@ -432,7 +421,7 @@ where
                 attestation_object: None,
             },
             authenticator_attachment: Some(self.authenticator().attachment_type()),
-            client_extension_results: AuthenticatorExtensionsClientOutputs::default(),
+            client_extension_results: AuthenticationExtensionsClientOutputs::default(),
         })
     }